# docs/agent_auth_openapi.sketch.yaml
openapi: 3.1.0
info:
  title: ZEXVRO Gate (Agent Auth)
  version: 0.1.0
  description: Dual-channel human/agent capability API. See docs/agent_auth_implementation_plan.md
servers:
  - url: http://localhost:4103
paths:
  /health:
    get:
      summary: Liveness
      responses:
        '200':
          description: OK
  /status:
    get:
      summary: Public capabilities (no secrets)
      responses:
        '200':
          description: OK
  /v1/challenges:
    post:
      summary: Start human or agent challenge
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [siteKey, action, channel, clientPublicKey]
              properties:
                siteKey: { type: string }
                action: { type: string }
                channel: { type: string, enum: [human, agent] }
                clientPublicKey: { type: string }
                origin: { type: string, format: uri }
                agentPublicKey: { type: string }
      responses:
        '201':
          description: Challenge created
        '403':
          description: Origin not allowed / agent not registered
  /v1/challenges/{id}/complete:
    post:
      summary: Complete ceremony and mint capability
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [siteKey, proof, proofType]
              properties:
                siteKey: { type: string }
                proof: { type: string }
                proofType: { type: string, enum: [soft_confirm, nonce_sign, webauthn] }
      responses:
        '200':
          description: Capability issued
  /v1/verify:
    post:
      summary: Server-side capability verification
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [capability, action, siteSecret]
              properties:
                capability: { type: string }
                action: { type: string }
                minClass: { type: string, enum: [human, agent, either] }
                expectedOrigin: { type: string, format: uri }
                siteSecret: { type: string }
      responses:
        '200':
          description: Valid
        '401':
          description: Invalid / mismatch
        '403':
          description: Policy or class denied
  /v1/admin/agents:
    post:
      summary: Register agent public key (auth TBD Cognito)
      responses:
        '201':
          description: Registered
components:
  headers:
    X-Zexvro-Capability:
      description: Short-lived capability JWT from Gate
      schema: { type: string }